restore the original kernel xdp_synproxy.c options
modified usage function always complain options not
specified correctly, restore the original usage from
kernel xdp_synproxy.c resolves the issue (reason unknown)

SYN flood test:

./xdppacket --iface ens192 --mss4 1460 --mss6 1440 --ports 80 --wscale 7 --ttl 64
....
libbpf: map 'jmp_syncookie_table': slot [0] set to prog 'syncookie_xdp' fd=93
Replacing allowed ports
Added port 80
Replacing TCP/IP options
SYNACKs generated: 0 (total 0)
SYNACKs generated: 1 (total 1)
SYNACKs generated: 1 (total 2)
SYNACKs generated: 1 (total 3)
SYNACKs generated: 1 (total 4)
SYNACKs generated: 1 (total 5)
SYNACKs generated: 1 (total 6)
SYNACKs generated: 1 (total 7)
SYNACKs generated: 1 (total 8)
SYNACKs generated: 1 (total 9)
SYNACKs generated: 1 (total 10)

from hping3:

hping3 <target> -S -y -p 80

note tcpdump on target ens192 interface should not see SYN packet
since it is dropped by XDP

DNS ratelimit test (DNS request rate limit 16 per second):

root@vincent-pc2:~# bpftool map dump id 754
[{
        "value": {
            ".data": [{
                    "ratelimit": 32
                },{
                    "numcpus": 2
                }
            ]
        }
    }
]

root@lxc-dev:~# while true; do dig @10.11.15.114 www.cttb.com +short ; sleep 0.01 ; done
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154
209.61.212.154

;; communications error to 10.11.15.114#53: timed out

+++ examples/c/xdppacket.bpf.c
@@ -54,7 +54,7 @@ struct meta_data {
 	__u16 unused;
 };
 
-static volatile __u16 ratelimit = 10;
+static volatile unsigned int ratelimit = 10;
 static volatile __u8 numcpus = 2;
 
 struct ipv4_key {
